# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Cue API
  version: 0.1.0
  description: |
    HTTP contract between cue-api and cue-ios. Source of truth for endpoints,
    request/response shapes, and status codes. See `docs/architecture.md` and
    `docs/specs/` for design rationale.
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.cue.app
    description: Production (placeholder — not yet deployed)
...
paths:
...
  /tasks:
    get:
      tags: [tasks]
      summary: List task occurrences in a calendar and date range
      description: |
        Returns the expanded occurrences of all tasks in `calendarId` that
        intersect the half-open window `[from, to)`. Each recurring anchor is
        expanded into one `OccurrenceDTO` per occurrence date; one-off tasks
        produce one `OccurrenceDTO` each. Exceptions (skips, overrides) are
        applied. Occurrence identity for per-instance actions is `(taskId, occurrenceStart)`.
      security:
        - bearerAuth: []
      parameters:
        - name: calendarId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: Calendar to read. Must belong to the authenticated user.
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date-time
          description: Inclusive lower bound of the occurrence window (ISO-8601).
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date-time
          description: Exclusive upper bound of the occurrence window (ISO-8601).
        - name: includeCompleted
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: When true, completed occurrences are included.
        - name: includeTodos
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: When true, timeless todos (no startAt) are included.
      responses:
        '200':
          description: Expanded occurrences for the window.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OccurrenceDTO'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Missing, invalid, or expired bearer token.
        '403':
          description: The calendar is owned by another user.
        '404':
          description: The calendar does not exist.
    post:
      tags: [tasks]
      summary: Create a task
      description: |
        Creates a task in the referenced calendar (which must belong to the
        authenticated user) and returns the persisted series row as a `TaskDTO`
        with the embedded recurrence config. Supports timed events, all-day events,
        timeless todos, recurring series (via `recurrence`), and group assignment.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDTO'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Missing, invalid, or expired bearer token.
        '403':
          description: The calendar is owned by another user.
        '404':
          description: The calendar does not exist.
  /tasks/search:
    get:
      tags: [tasks]
      summary: Search tasks by text
      description: |
        Case-insensitive ILIKE search over `title` + `notes` across ALL the authed
        user's calendars (M1 — NOT window-bound). `q` is required and trimmed; an
        optional `groupId` narrows to one group; `limit` defaults to 25 and is
        clamped to 50. Returns the matched series rows (recurring matches return the
        anchor) as `TaskSearchResultDTO[]`.
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Case-insensitive search term matched against title + notes (trimmed).
        - name: groupId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Narrow the search to a single task group.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 25
          description: Maximum matches to return (clamped to 50).
      responses:
        '200':
          description: Matching tasks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskSearchResultDTO'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Missing, invalid, or expired bearer token.
  /tasks/daily-counts:
    get:
      tags: [tasks]
      summary: Per-day task occurrence counts in a calendar and date range
      description: |
        Returns the number of task occurrences bucketed onto each local date
        (`YYYY-MM-DD`, in each task's timezone) within the half-open window
        `[from, to)` for `calendarId`. Occurrence-aware: each recurring anchor is
        expanded (exceptions applied) and every resulting occurrence contributes
        to its date bucket — a weekly task with 4 occurrences in range adds 4
        across the respective days. Counts on the same local date are summed and
        zero-count days are omitted from the response.
      security:
        - bearerAuth: []
      parameters:
        - name: calendarId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: Calendar to count for. Must belong to the authenticated user.
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date-time
          description: Inclusive lower bound of the window (ISO-8601).
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date-time
          description: Exclusive upper bound of the window (ISO-8601).
        - name: includeCompleted
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: When true, completed occurrences are counted.
      responses:
        '200':
          description: Per-day occurrence counts for the window.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyCountsDTO'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Missing, invalid, or expired bearer token.
        '403':
          description: The calendar is owned by another user.
        '404':
          description: The calendar does not exist.
  /tasks/changes:
    get:
      tags: [tasks]
      summary: Precise change-delta for a calendar since an opaque cursor
      description: |
        Returns everything that changed in `calendarId` since the opaque cursor
        `since` (the `serverTime` of the previous response, echoed verbatim):

        - `tasks` — changed series rows (`updatedAt > since`), each with the
          embedded `recurrence` config.
        - `deleted` — ids of series soft-deleted (`deletedAt > since`); visible
          via `withDeleted`.
        - `exceptions` — per-occurrence overrides changed since the cursor.
          Exceptions are hard-deleted (no tombstone), so a *removed* exception is
          surfaced indirectly by the parent Task's `updatedAt` bump (the write
          path touches the parent).
        - `serverTime` — the new opaque cursor (server clock, not the client's),
          captured BEFORE the reads so no change is missed between read and stamp.

        On the FIRST call, omit `since`: the response carries empty sets and only
        the cursor for the client to store, and the client falls back to a full
        per-window SWR sync for the initial load. The cursor is opaque — today an
        ISO-8601 server time, internally swappable for a monotonic syncVersion
        behind the same shape.
      security:
        - bearerAuth: []
      parameters:
        - name: calendarId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: Calendar to diff. Must belong to the authenticated user.
        - name: since
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: |
            Opaque cursor — the previous response's `serverTime`, echoed
            verbatim. Omit on the first call to force full per-window SWR.
        - name: includeTodos
          in: query
          required: false
          schema:
            type: boolean
            default: true
          description: When false, timeless todos are excluded from the changed set.
      responses:
        '200':
          description: The change-delta since the cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangesDTO'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Missing, invalid, or expired bearer token.
        '403':
          description: The calendar is owned by another user.
        '404':
          description: The calendar does not exist.
  /tasks/{id}:
    get:
      tags: [tasks]
      summary: Get a task series row
      description: |
        Returns the full `TaskDTO` for the series anchor, including the embedded
        `recurrence` rule ready for an edit form. Use this to open the detail /
        edit screen.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Task id.
      responses:
        '200':
          description: Task series row.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDTO'
        '401':
          description: Missing, invalid, or expired bearer token.
        '403':
          description: Task belongs to another user.
        '404':
          description: No task with that id.
    patch:
      tags: [tasks]
      summary: Update a task (all / one-off scope)
      description: |
        Updates the task's fields in the "all" scope — the whole series or the
        one-off task. Omitted fields are left unchanged. `recurrence: null` clears
        the inline config; a `RecurrenceRuleInput` replaces or adds it. `color` and
        `requiresCompletion` may likewise be set or cleared (null re-inherits the
        group). Returns the updated `TaskDTO`.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Task id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Updated task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDTO'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Missing, invalid, or expired bearer token.
        '403':
          description: Task belongs to another user.
        '404':
          description: No task with that id.
    delete:
      tags: [tasks]
      summary: Delete a task
      description: |
        Soft-deletes the task (and its series, for recurring tasks). Returns
        `{ id }` with HTTP 200 — NOT 204, so the iOS client can decode the body.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Task id.
      responses:
        '200':
          description: Task deleted.
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id:
                    type: string
                    format: uuid
        '401':
          description: Missing, invalid, or expired bearer token.
        '403':
          description: Task belongs to another user.
        '404':
          description: No task with that id.
  /tasks/{id}/completion:
    patch:
      tags: [tasks]
      summary: Toggle task or occurrence completion
      description: |
        When `occurrenceStart` is provided, toggles completion for that single
        occurrence of a recurring task (per-instance). When omitted, toggles the
        series anchor or one-off task. Returns a `CompletionResultDTO`.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Task id (series anchor or one-off).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetCompletionRequest'
      responses:
        '200':
          description: Completion state updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompletionResultDTO'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Missing, invalid, or expired bearer token.
        '403':
          description: Task belongs to another user.
        '404':
          description: No task with that id.
  /tasks/{id}/skip:
    post:
      tags: [tasks]
      summary: Skip a single recurring occurrence
      description: |
        Marks a single occurrence as skipped by upserting a
        `TaskOccurrenceException` with `isSkipped: true`. The occurrence
        disappears from the calendar without affecting other instances or the
        series rule. Returns `{ ok: true }`.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Task id (series anchor).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkipOccurrenceRequest'
      responses:
        '200':
          description: Occurrence skipped.
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
                    enum: [true]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Missing, invalid, or expired bearer token.
        '403':
          description: Task belongs to another user.
        '404':
          description: No task with that id, or occurrenceStart is not a real occurrence.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TaskDTO:
      type: object
      description: |
        A task series row. Returned by POST /tasks, PATCH /tasks/:id, GET /tasks/:id.
        Includes the embedded recurrence config for the edit form.
      required:
        [id, calendarId, groupId, title, notes, startAt, endAt, isAllDay, timezone,
         requiresCompletion, color, icon, reminders, completedAt, recurrence,
         notificationStrategyId, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        calendarId:
          type: string
          format: uuid
        groupId:
          type: [string, 'null']
          format: uuid
        title:
          type: string
        notes:
          type: [string, 'null']
        icon:
          type: [string, 'null']
          description: Per-task icon (e.g. an SF Symbol name); null when iconless. (S4)
        reminders:
          type: array
          items:
            $ref: '#/components/schemas/ReminderResultDTO'
          description: |
            The task's own per-task reminders (offset + channel), earliest first.
            Empty when the task carries no per-task reminders. (S1)
        startAt:
          type: [string, 'null']
          format: date-time
          description: ISO-8601 start; null for a timeless todo.
        endAt:
          type: [string, 'null']
          format: date-time
        isAllDay:
          type: boolean
        timezone:
          type: string
          description: IANA zone — the anchor for recurrence/DST expansion.
        requiresCompletion:
          type: [boolean, 'null']
          description: |
            Per-task override; null inherits the group, then defaults to false.
            See the effective-settings resolver (ADR 0054, task-wins).
        color:
          type: [string, 'null']
          description: |
            A TaskColor preset name OR a #RRGGBB hex; null inherits the group.
            Effective color resolves task-wins (ADR 0054).
        completedAt:
          type: [string, 'null']
          format: date-time
        recurrence:
          oneOf:
            - $ref: '#/components/schemas/RecurrenceConfig'
            - type: 'null'
          description: |
            Inline recurrence config (JSONB) for the edit form; null when not
            recurring. ADR 0054 — no separate rule id.
        notificationStrategyId:
          type: [string, 'null']
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OccurrenceDTO:
      type: object
      description: |
        A single expanded occurrence. Returned by GET /tasks. One per visible
        calendar cell. Stable identity = (taskId, occurrenceStart).
      required:
        [taskId, calendarId, groupId, groupColorHex, originalStart, occurrenceStart, occurrenceEnd,
         title, notes, isAllDay, timezone, requiresCompletion, color, icon, completedAt,
         isRecurring, isException]
      properties:
        taskId:
          type: string
          format: uuid
          description: Series anchor id — same across all occurrences of a recurring task.
        calendarId:
          type: string
          format: uuid
        groupId:
          type: [string, 'null']
          format: uuid
        groupColorHex:
          type: [string, 'null']
          example: '#E27921'
          description: The owning group's color (preset name or #RRGGBB); null when ungrouped.
        originalStart:
          type: [string, 'null']
          format: date-time
          description: Rule-generated start BEFORE any override; stable instance key WITH taskId.
        occurrenceStart:
          type: [string, 'null']
          format: date-time
          description: Effective start with any overrideStartAt applied.
        occurrenceEnd:
          type: [string, 'null']
          format: date-time
        title:
          type: string
        notes:
          type: [string, 'null']
        isAllDay:
          type: boolean
        timezone:
          type: string
        requiresCompletion:
          type: boolean
          description: |
            Effective value (task ?? group ?? false) per the ADR 0054 resolver,
            already resolved for the rendered occurrence.
        color:
          type: [string, 'null']
          description: |
            Effective color (task ?? group) per the ADR 0054 resolver — a TaskColor
            preset name OR a #RRGGBB hex; null when neither task nor group sets one.
        icon:
          type: [string, 'null']
          description: Per-task icon (e.g. an SF Symbol name); null when iconless. (S4)
        completedAt:
          type: [string, 'null']
          format: date-time
          description: Per-instance completion for recurring tasks.
        isRecurring:
          type: boolean
        isException:
          type: boolean
    DailyCountsDTO:
      type: object
      description: |
        Per-day occurrence counts. Returned by GET /tasks/daily-counts. The
        `counts` map is keyed by local date (YYYY-MM-DD, in each task's
        timezone); zero-count days are omitted.
      required: [counts]
      properties:
        counts:
          type: object
          additionalProperties:
            type: integer
          description: Occurrence count per local date (YYYY-MM-DD).
          example:
            '2026-06-20': 3
            '2026-06-21': 1
            '2026-06-23': 2
...
    CreateTaskRequest:
      type: object
      required: [calendarId, title, timezone]
      properties:
        calendarId:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
        notes:
          type: [string, 'null']
        startAt:
          type: [string, 'null']
          format: date-time
        endAt:
          type: [string, 'null']
          format: date-time
        isAllDay:
          type: boolean
          default: false
        timezone:
          type: string
          description: IANA zone.
        requiresCompletion:
          type: [boolean, 'null']
          description: |
            Omit / null to inherit the group, falling back to false (ADR 0054
            resolver). No column default.
        color:
          type: [string, 'null']
          description: A TaskColor preset name OR a #RRGGBB hex; null/omit inherits the group.
        icon:
          type: [string, 'null']
          description: Optional per-task icon (e.g. an SF Symbol name). (S4)
        reminders:
          type: array
          items:
            $ref: '#/components/schemas/ReminderRequest'
          description: Per-task reminders (offset + channel) created with the task. (S1)
        groupId:
          type: string
          format: uuid
          description: Optional group (must live in the same calendar).
        recurrence:
          $ref: '#/components/schemas/RecurrenceRuleInput'
...
    SkipOccurrenceRequest:
      type: object
      required: [occurrenceStart]
      properties:
        occurrenceStart:
          type: string
          format: date-time
          description: ISO-8601 occurrence to skip.
...
    TaskSearchResultDTO:
      type: object
      description: |
        A single task search hit (M1). Carries the matched series row's identity,
        its owning group's id + color, and its anchor date/time — for a recurring
        series, the anchor occurrence (the calendar view expands occurrences
        separately).
      required:
        [taskId, calendarId, groupId, groupColorHex, title, notes, startAt, endAt,
         isAllDay, timezone, isRecurring]
      properties:
        taskId:
          type: string
          format: uuid
        calendarId:
          type: string
          format: uuid
        groupId:
          type: [string, 'null']
          format: uuid
        groupColorHex:
          type: [string, 'null']
          example: '#E27921'
          description: The owning group's color (preset name or #RRGGBB); null when ungrouped.
        title:
          type: string
        notes:
          type: [string, 'null']
        startAt:
          type: [string, 'null']
          format: date-time
        endAt:
          type: [string, 'null']
          format: date-time
        isAllDay:
          type: boolean
        timezone:
          type: string
        isRecurring:
          type: boolean
          description: True when the matched task is a recurring series (startAt is its anchor).
...
  responses:
    ValidationError:
      description: Request DTO failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorBody'
...